+++ package.json
@@ -10,6 +10,7 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.23",
     "@types/react-dom": "^16.9.5",
+    "@types/react-router": "^5.1.4",
     "@types/react-router-dom": "^5.1.3",
     "@types/styled-components": "^5.0.1",
     "react": "^16.13.0",
