+++ src/Components/Layout/Layout.tsx
@@ -1,0 +1,47 @@
+import React from 'react';
+import Nav from 'Components/Nav/Nav';
+import styled from 'styled-components';
+
+interface IProps {
+  children: React.ReactNode;
+}
+const Layout = ({ children }: IProps) => {
+  return (
+    <TotalBackground>
+      <Nav />
+      <MainContainer>{children}</MainContainer>
+    </TotalBackground>
+  );
+};
+
+export default Layout;
+
+const TotalBackground = styled.div`
+  display: flex;
+  overflow-x: hidden;
+  ${({ theme }) => theme.media.desktop`
+ display: flex;
+  overflow-x: hidden;
+  `}
+`;
+
+const MainContainer = styled.div`
+  color: black;
+  margin-top: 67px;
+  min-height: 600px;
+  background-color: #fbfbfb;
+  width: 100%;
+
+  ${({ theme }) => theme.media.desktop`
+ margin-top:0;
+ height: 100vh;
+  color: black;
+  padding-bottom: 245px;
+  min-height: 600px;
+  padding-left: 225px;
+  background-color: #fbfbfb;
+  width: 100%;
+    display: flex;
+  justify-content: center;
+  `}
+`;
