+++ src/Pages/DjStation/DjStation.tsx
@@ -1,12 +1,8 @@
 import React, { Component } from 'react';
+import Layout from 'Components/Layout/Layout';
 import ThemeIcon from 'Img/themeIcon.png';
 // import StationComponent from 'Components/StationComponent';
 import styled from 'styled-components';
-
-interface IProps {
-  props: any;
-  text: string;
-}
 
 export default class DjStation extends Component {
   //  constructor(props) {
@@ -28,107 +24,30 @@ export default class DjStation extends Component {
 
   render() {
     return (
-      <TotalBackground>
-        {/* ************nav************* */}
-        <Nav></Nav>
-        {/* ********nav 끝******** */}
-        <MainContainer>
-          <MainContent>
-            {/* 메인 글씨 랑 테마 버튼 시작************** */}
-            <MainText>
-              DJ 스테이션{' '}
-              <ThemeBox>
-                <ThemeIconBox src={ThemeIcon}></ThemeIconBox>테마
-              </ThemeBox>
-            </MainText>
-            {/* 메인 글씨랑 테마 버튼 끝 ******************* */}
-            {/* 느낌 별 스테이션 시작 *******************8 */}
-            <StationComponentContainer>
-              <StationComponentContainerText>
-                느낌별 스테이션
-              </StationComponentContainerText>
-              <StationWrap>
-                <StaionWrapUl></StaionWrapUl>
-              </StationWrap>
-            </StationComponentContainer>
-          </MainContent>
-        </MainContainer>
-      </TotalBackground>
+      <Layout>
+        <MainContent>
+          {/* 메인 글씨 랑 테마 버튼 시작************** */}
+          <MainText>
+            DJ 스테이션{' '}
+            <ThemeBox>
+              <ThemeIconBox src={ThemeIcon}></ThemeIconBox>테마
+            </ThemeBox>
+          </MainText>
+          {/* 메인 글씨랑 테마 버튼 끝 ******************* */}
+          {/* 느낌 별 스테이션 시작 *******************8 */}
+          <StationComponentContainer>
+            <StationComponentContainerText>
+              느낌별 스테이션
+            </StationComponentContainerText>
+            <StationWrap>
+              <StaionWrapUl></StaionWrapUl>
+            </StationWrap>
+          </StationComponentContainer>
+        </MainContent>
+      </Layout>
     );
   }
 }
-const TotalBackground = styled.div`
-  display: flex;
-  overflow-x: hidden;
-  ${({ theme }) => theme.media.desktop`
- display: flex;
-  overflow-x: hidden;
-  `}
-  ${({ theme }) => theme.media.tablet`
-  display: flex;
-  overflow-x: hidden;
-  `}
-`;
-const Nav = styled.div`
-/* 제일 작을 때, 768px 이하**************** */
-  width: 100%;
-  height: 67px;
-  background-color: black;
-  position: fixed;
-  top: 0;
-  z-index: 10;
-  /* 1100정도 이상********************* */
-  ${({ theme }) => theme.media.desktop`
-background-color: black;
- width: 225px;
-  height: 100vh;
-  position: fixed;
-  top: 0;
-  ${(props: IProps) => `$::before{content:"데스크톱 ${props.text}"}`};
-  `}
-  /* 이거 밑에 768~1100정도까지********************************8 */
-  ${({ theme }) => theme.media.tablet`
-  background-color:gray;
-   width: 225px;
-  height: 100vh;
-  position: fixed;
-  top: 0;
-  ${(props: IProps) => `&::before{content:"태블릿${props.text}"}`};
-  `}
-`;
-const MainContainer = styled.div`
-  color: black;
-  margin-top: 67px;
-  min-height: 600px;
-  background-color: #fbfbfb;
-  width: 100%;
-
-  ${({ theme }) => theme.media.desktop`
- margin-top:0;
- height: 100vh;
-  color: black;
-  padding-bottom: 245px;
-  min-height: 600px;
-  padding-left: 225px;
-  background-color: #fbfbfb;
-  width: 100%;
-    display: flex;
-  justify-content: center;
-  ${(props: IProps) => `$::before{content:"데스크톱 ${props.text}"}`};
-  `}
-  ${({ theme }) => theme.media.tablet`
-  margin-top:0;
- height: 100vh;
-  color: black;
-  padding-bottom: 245px;
-  min-height: 600px;
-  padding-left: 225px;
-  background-color: #fbfbfb;
-  width: 100%;
- 
-  ${(props: IProps) => `&::before{content:"태블릿${props.text}"}`};
-  `}
-`;
 
 const MainContent = styled.div`
   margin: 0 43px;
@@ -136,13 +55,7 @@ const MainContent = styled.div`
   padding-top : 20px;
   margin : 0 43px;
   width : 964px;
-  ${(props: IProps) => `$::before{content:"데스크톱 ${props.text}"}`};
-  `}
-  ${({ theme }) => theme.media.tablet`
-  padding-top: 61px;
-  margin: 0 43px;
-  width: auto;
-  ${(props: IProps) => `&::before{content:"태블릿${props.text}"}`};
+
   `}
 `;
 const ThemeBox = styled.button`
